fix: AUT-4485 extract metadata list compatibility checker
